Crate kong

                              )                 
                           ( /(          (  (   
                            )\())(   (    )\))(  
                           ((_)\ )\  )\ )((_))\  
                           | |(_|(_)_(_/( (()(_)
                           | / / _ \ ' \)) _` |  
                           |_\_\___/_||_|\__, |  
                         secure web node |___/ v0.1.0

Re-exports

pub use krypto;

pub use rouille as server;

Modules

defaults

ğŸ® kong default, constant values

inputs

âŒ¨ï¸ kong user input management

log

ğŸ“‡ kong node logging

validate

ğŸ”¬ kong user input validation

Macros

json

Construct a serde_json::Value from a JSON literal.

Structs

ErrorResponse

ğŸ´ API error response

JsonError

This type represents all possible errors that can occur when serializing or deserializing JSON data.

Konfig

ğŸ›ï¸ Server configuration

Kong

ğŸ”¥ Kong object

Enums

JsonValue

Represents any valid JSON value.

KError

ğŸš¨ kong error management

Method

ğŸš¥ HTTP methods

Traits

Kontrol

ğŸ® API Enpoint kontrollers

Functions

json_from_str

Deserialize an instance of type T from a string of JSON text.

kroute

ğŸŒ€ kong request routing